refactor all extraction utils
