+++ packages/plugin-autocapture-browser/src/frustration-plugin.ts
@@ -10,17 +10,12 @@ import {
 } from '@amplitude/analytics-core';
 import * as constants from './constants';
 import { fromEvent, map, Observable, Subscription, share } from 'rxjs';
-import {
-  addAdditionalEventProperties,
-  createShouldTrackEvent,
-  ElementBasedTimestampedEvent,
-  getEventProperties,
-  NavigateEvent,
-} from './helpers';
+import { createShouldTrackEvent, ElementBasedTimestampedEvent, NavigateEvent } from './helpers';
 import { trackDeadClick } from './autocapture/track-dead-click';
 import { trackRageClicks } from './autocapture/track-rage-click';
 import { AllWindowObservables, ObservablesEnum } from './autocapture-plugin';
 import { createClickObservable, createMutationObservable } from './observables';
+import { DataExtractor } from './data-extractor';
 
 type BrowserEnrichmentPlugin = EnrichmentPlugin<BrowserClient, BrowserConfig>;
 
@@ -35,6 +30,8 @@ export const frustrationPlugin = (options: FrustrationInteractionsOptions = {}):
 
   const dataAttributePrefix = options.dataAttributePrefix ?? DEFAULT_DATA_ATTRIBUTE_PREFIX;
 
+  const dataExtractor = new DataExtractor(options);
+
   // combine the two selector lists to determine which clicked elements should be filtered
   const combinedCssSelectors = [...new Set([...rageCssSelectors, ...deadCssSelectors])];
 
@@ -43,7 +40,7 @@ export const frustrationPlugin = (options: FrustrationInteractionsOptions = {}):
     // Create Observables from direct user events
     const clickObservable = createClickObservable('pointerdown').pipe(
       map((click) => {
-        return addAdditionalEventProperties(
+        return dataExtractor.addAdditionalEventProperties(
           click,
           'click',
           combinedCssSelectors,
@@ -60,7 +57,7 @@ export const frustrationPlugin = (options: FrustrationInteractionsOptions = {}):
     if (window.navigation) {
       navigateObservable = fromEvent<NavigateEvent>(window.navigation, 'navigate').pipe(
         map((navigate) =>
-          addAdditionalEventProperties(navigate, 'navigate', combinedCssSelectors, dataAttributePrefix),
+          dataExtractor.addAdditionalEventProperties(navigate, 'navigate', combinedCssSelectors, dataAttributePrefix),
         ),
         share(),
       );
@@ -68,7 +65,9 @@ export const frustrationPlugin = (options: FrustrationInteractionsOptions = {}):
 
     // Track DOM Mutations
     const enrichedMutationObservable = createMutationObservable().pipe(
-      map((mutation) => addAdditionalEventProperties(mutation, 'mutation', combinedCssSelectors, dataAttributePrefix)),
+      map((mutation) =>
+        dataExtractor.addAdditionalEventProperties(mutation, 'mutation', combinedCssSelectors, dataAttributePrefix),
+      ),
       share(),
     );
 
@@ -104,7 +103,8 @@ export const frustrationPlugin = (options: FrustrationInteractionsOptions = {}):
     const deadClickSubscription = trackDeadClick({
       amplitude,
       allObservables,
-      getEventProperties: (actionType, element) => getEventProperties(actionType, element, dataAttributePrefix),
+      getEventProperties: (actionType, element) =>
+        dataExtractor.getEventProperties(actionType, element, dataAttributePrefix),
       shouldTrackDeadClick,
     });
     subscriptions.push(deadClickSubscription);
