+++ packages/plugin-autocapture-browser/src/libs/messenger.ts
@@ -9,7 +9,7 @@ import {
 import { asyncLoadScript, generateUniqueId } from '../helpers';
 import { ILogger, Messenger, ActionType } from '@amplitude/analytics-core';
 import { VERSION } from '../version';
-import type { DataExtractor } from '../data-extractor';
+import { DataExtractor } from '../data-extractor';
 
 export type Action =
   | 'ping'
@@ -133,21 +133,23 @@ export class WindowMessenger implements Messenger {
     delete this.requestCallbacks[response.id];
   }
 
-  setup({
-    logger,
-    endpoint,
-    isElementSelectable,
-    cssSelectorAllowlist,
-    actionClickAllowlist,
-    dataExtractor,
-  }: {
-    logger?: ILogger;
-    endpoint?: string;
-    isElementSelectable?: (action: InitializeVisualTaggingSelectorData['actionType'], element: Element) => boolean;
-    cssSelectorAllowlist?: string[];
-    actionClickAllowlist?: string[];
-    dataExtractor: DataExtractor;
-  } = {}) {
+  setup(
+    {
+      logger,
+      endpoint,
+      isElementSelectable,
+      cssSelectorAllowlist,
+      actionClickAllowlist,
+      dataExtractor,
+    }: {
+      logger?: ILogger;
+      endpoint?: string;
+      isElementSelectable?: (action: InitializeVisualTaggingSelectorData['actionType'], element: Element) => boolean;
+      cssSelectorAllowlist?: string[];
+      actionClickAllowlist?: string[];
+      dataExtractor: DataExtractor;
+    } = { dataExtractor: new DataExtractor({}) },
+  ) {
     this.logger = logger;
     // If endpoint is customized, don't override it.
     if (endpoint && this.endpoint === AMPLITUDE_ORIGIN) {
